Reorg of domain toolkit
